Modify SearchPictureComponent html template to place the infiniteSCrollDirective on a div with an auto overflow. So that the scroll listener host in the directive has the div as a target.
Added code in the onScroll method for the signal to be fired
only when the scrollbar is at the bottom.
This way signal is emitted, thus http request is fired only
when new images have to be loaded.

Next I will try to remove the button search and make the keyword appear
in the uri path